# docs/swagger.yaml
basePath: /api/v1
definitions:
  gopet_internal_core_transport_response.ErorrResponse:
    properties:
      error:
        example: full error text
        type: string
      message:
        example: short human-readable message
        type: string
    type: object
...
  internal_features_tasks_transport_http.CreateTaskRequest:
    properties:
      author_user_id:
        example: 123
        type: integer
      description:
        example: i gotta have a math lesson tomorrow so i need to do my hw
        maxLength: 1000
        minLength: 1
        type: string
      title:
        example: make my math homework
        maxLength: 100
        minLength: 1
        type: string
    required:
    - author_user_id
    - title
    type: object
  internal_features_tasks_transport_http.CreateTaskResponse:
    properties:
      author_user_id:
        example: 5
        type: integer
      completed:
        example: false
        type: boolean
      completed_at:
        example: "null"
        type: string
      created_at:
        example: "2026-02-26T10:30:00Z"
        type: string
      description:
        example: i need to do my homework
        type: string
      id:
        example: 123
        type: integer
      title:
        example: homework
        type: string
      version:
        example: 12
        type: integer
    type: object
  internal_features_tasks_transport_http.GetTaskResponse:
    properties:
      author_user_id:
        example: 5
        type: integer
      completed:
        example: false
        type: boolean
      completed_at:
        example: "null"
        type: string
      created_at:
        example: "2026-02-26T10:30:00Z"
        type: string
      description:
        example: i need to do my homework
        type: string
      id:
        example: 123
        type: integer
      title:
        example: homework
        type: string
      version:
        example: 12
        type: integer
    type: object
  internal_features_tasks_transport_http.PatchTaskRequest:
    properties:
      completed:
        example: false
        type: boolean
      description:
        example: i need to make hw
        type: string
      title:
        example: Make HW
        type: string
    type: object
  internal_features_tasks_transport_http.PatchUserResponse:
    properties:
      author_user_id:
        example: 5
        type: integer
      completed:
        example: false
        type: boolean
      completed_at:
        example: "null"
        type: string
      created_at:
        example: "2026-02-26T10:30:00Z"
        type: string
      description:
        example: i need to do my homework
        type: string
      id:
        example: 123
        type: integer
      title:
        example: homework
        type: string
      version:
        example: 12
        type: integer
    type: object
  internal_features_tasks_transport_http.TaskDTOResponse:
    properties:
      author_user_id:
        example: 5
        type: integer
      completed:
        example: false
        type: boolean
      completed_at:
        example: "null"
        type: string
      created_at:
        example: "2026-02-26T10:30:00Z"
        type: string
      description:
        example: i need to do my homework
        type: string
      id:
        example: 123
        type: integer
      title:
        example: homework
        type: string
      version:
        example: 12
        type: integer
    type: object
  internal_features_users_transort_http.CreateUserRequest:
    properties:
      full_name:
        example: Ivan Ivanon
        maxLength: 100
        minLength: 3
        type: string
      phone_number:
        example: "+38097123804"
        maxLength: 15
        minLength: 10
        type: string
    required:
    - full_name
    type: object
  internal_features_users_transort_http.CreateUserResponse:
    properties:
      full_name:
        example: Ivan Ivanov
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+38097123804"
        type: string
      version:
        example: 2
        type: integer
    type: object
...
  internal_features_users_transort_http.UserDTOResponse:
    properties:
      full_name:
        example: Ivan Ivanov
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+38097123804"
        type: string
      version:
        example: 2
        type: integer
    type: object
host: 172.25.87.4:5050
info:
  contact: {}
  description: gopet rest api schema
  title: Golang GOPET API
  version: "1.0"
paths:
...
  /tasks:
    get:
      description: Get tasks in system
      parameters:
      - description: Filtraion by user id
        in: query
        name: user_id
        type: string
      - description: Pagination limit parameter
        in: query
        name: limit
        type: integer
      - description: Pagination offset parameter
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved tasks
          schema:
            items:
              $ref: '#/definitions/internal_features_tasks_transport_http.TaskDTOResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gopet_internal_core_transport_response.ErorrResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gopet_internal_core_transport_response.ErorrResponse'
      summary: Get task
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Create new task in system
      parameters:
      - description: CreateTask request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_tasks_transport_http.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created task
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.CreateTaskResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gopet_internal_core_transport_response.ErorrResponse'
        "404":
          description: Author not found
          schema:
            $ref: '#/definitions/gopet_internal_core_transport_response.ErorrResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gopet_internal_core_transport_response.ErorrResponse'
      summary: Create task
      tags:
      - tasks
  /tasks/{id}:
    delete:
      description: Delete task in system
      parameters:
      - description: ID of the task to be deleted
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Successfully deleted task
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gopet_internal_core_transport_response.ErorrResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/gopet_internal_core_transport_response.ErorrResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gopet_internal_core_transport_response.ErorrResponse'
      summary: Delete task
      tags:
      - tasks
    get:
      description: Get task in system
      parameters:
      - description: ID of the task to be retrieved
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved task
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.GetTaskResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gopet_internal_core_transport_response.ErorrResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/gopet_internal_core_transport_response.ErorrResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gopet_internal_core_transport_response.ErorrResponse'
      summary: Get task
      tags:
      - tasks
    patch:
      consumes:
      - application/json
      description: |-
        Get tasks in system
        ### Three-state logic:
        1. **Field is not provided:** nothing to be done
        2. **Field is provided("go for a walk"):** redact
        3. **Field is provided(null):** set to null
      parameters:
      - description: ID of the task to be patched
        in: path
        name: id
        required: true
        type: integer
      - description: request body of PatchTask
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_tasks_transport_http.PatchTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully patched task
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.PatchUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gopet_internal_core_transport_response.ErorrResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/gopet_internal_core_transport_response.ErorrResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/gopet_internal_core_transport_response.ErorrResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gopet_internal_core_transport_response.ErorrResponse'
      summary: Patch task
      tags:
      - tasks
  /users:
    get:
      description: Get existing users in system with optional pagination
      parameters:
      - description: Pagination limit parameter
        in: query
        name: limit
        type: integer
      - description: Pagination offset parameter
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved users
          schema:
            items:
              $ref: '#/definitions/internal_features_users_transort_http.UserDTOResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gopet_internal_core_transport_response.ErorrResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gopet_internal_core_transport_response.ErorrResponse'
      summary: Get Users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create new user in system
      parameters:
      - description: CreateUser request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_users_transort_http.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created user
          schema:
            $ref: '#/definitions/internal_features_users_transort_http.CreateUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gopet_internal_core_transport_response.ErorrResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gopet_internal_core_transport_response.ErorrResponse'
      summary: Create User
      tags:
      - users
...
swagger: "2.0"