# internal/features/users/transort/http/dto_common.go
package users_transport_http

import "gopet/internal/core/domain"

type UserDTOResponse struct {
	ID          int     `json:"id"          example:"10"`
	Version     int     `json:"version"     example:"2"`
	FullName    string  `json:"full_name"    example:"Ivan Ivanov"`
	PhoneNumber *string `json:"phone_number" example:"+38097123804"`
}

func userDTOFromDomain(user domain.User) UserDTOResponse {
	return UserDTOResponse{
		ID:          user.ID,
		Version:     user.Version,
		FullName:    user.FullName,
		PhoneNumber: user.PhoneNumber,
	}
}

func usersDTOFromDoamins(users []domain.User) []UserDTOResponse {
	usersDTO := make([]UserDTOResponse, len(users))

	for i, user := range users {
		usersDTO[i] = userDTOFromDomain(user)
	}
	return usersDTO
}